[RSDK-6525] Have the RTK components manage their own data (#3680)
...without relying on an NMEA component inside them.

Here are the main changes:
- The `rtkutils` package has been renamed `gpsutils`
- The `DataReader` interface has been moved from the `gpsnmea` package to `gpsutils`. 
- `newI2cDataReader()` and `newSerialDataReader()` take different arguments, namely a `I2CConfig` or `SerialConfig` instead of raw values taken from within them. They're in charge of setting their own default values if any are missing, which is how I should have written it originally.
- `NewCachedData()` takes 2 new arguments: a `DataReader` and a logger. 
- The `CachedData` struct spawns its own managed goroutine to read its `DataReader`, and shuts it down when you `Close()` it. The goroutine that used to do this within `NMEAMovementSensor` has been removed.
- The `NmeaMovementSensor` interface no longer has a `Start()` function. That function should have been private from the beginning, and has now been moved inside `CachedData` anyway.
- The `rtkSerial` and `rtkI2C` contain an gpsutils.CachedData` instead of a `MovementSensor` component. 
- Unit tests for `rtkSerial` and `rtkI2C` that no longer compile because they mocked out an NMEA component to give to the RTK component have been deleted. As Susmita says, we can figure out the right unit tests once the refactoring is over.
- In the `rtkSerial` struct, if `connectToNTRIP()` encounters an error and is supposed to return an error, it returns _that_ one instead of whatever is stored within `g.err`. Without this, we could encounter errors and then return nil, as though we did not encounter errors, leading to panicking from a nil pointer dereference because it looked like the function succeeded even though it didn't.

Tried on piworld: all the different sensors (nmea-serial, rtk-i2c, rtk-serial, and dual) all eventually start up and give believable readings. Some of them take a long time to start up, as though they're making requests to web servers in their constructors or something. but that's a problem for another day/ticket.
