+++ components/movementsensor/gpsutils/cachedData.go
@@ -1,4 +1,4 @@
-package rtkutils
+package gpsutils
 
 import (
 	"context"
@@ -8,13 +8,22 @@ import (
 	"github.com/golang/geo/r3"
 	geo "github.com/kellydunn/golang-geo"
 	"github.com/pkg/errors"
+	goutils "go.viam.com/utils"
 
 	"go.viam.com/rdk/components/movementsensor"
+	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/spatialmath"
 	"go.viam.com/rdk/utils"
 )
 
 var errNilLocation = errors.New("nil gps location, check nmea message parsing")
+
+// DataReader represents a way to get data from a GPS NMEA device. We can read data from it using
+// the channel in Messages, and we can close the device when we're done.
+type DataReader interface {
+	Messages() chan string
+	Close() error
+}
 
 // CachedData allows the use of any MovementSensor chip via a DataReader.
 type CachedData struct {
@@ -24,15 +33,62 @@ type CachedData struct {
 	err                movementsensor.LastError
 	lastPosition       movementsensor.LastPosition
 	lastCompassHeading movementsensor.LastCompassHeading
+
+	dev    DataReader
+	logger logging.Logger
+
+	cancelCtx               context.Context
+	cancelFunc              func()
+	activeBackgroundWorkers sync.WaitGroup
 }
 
 // NewCachedData creates a new CachedData object.
-func NewCachedData() CachedData {
-	return CachedData{
+func NewCachedData(dev DataReader, logger logging.Logger) *CachedData {
+	cancelCtx, cancelFunc := context.WithCancel(context.Background())
+	g := CachedData{
 		err:                movementsensor.NewLastError(1, 1),
 		lastPosition:       movementsensor.NewLastPosition(),
 		lastCompassHeading: movementsensor.NewLastCompassHeading(),
+		dev:                dev,
+		logger:             logger,
+		cancelCtx:          cancelCtx,
+		cancelFunc:         cancelFunc,
 	}
+	g.start()
+	return &g
+}
+
+// start begins reading nmea messages from dev and updates gps data.
+func (g *CachedData) start() {
+	g.activeBackgroundWorkers.Add(1)
+	goutils.PanicCapturingGo(func() {
+		defer g.activeBackgroundWorkers.Done()
+
+		messages := g.dev.Messages()
+		done := g.cancelCtx.Done()
+		for {
+			// First, check if we're supposed to shut down.
+			select {
+			case <-done:
+				return
+			default:
+			}
+
+			// Next, wait until either we're supposed to shut down or we have new data to process.
+			select {
+			case <-done:
+				return
+			case message := <-messages:
+				// Update our struct's gps data in-place
+				err := g.ParseAndUpdate(message)
+				if err != nil {
+					g.logger.CWarnf(g.cancelCtx, "can't parse nmea sentence: %#v", err)
+					g.logger.Debug("Check: GPS requires clear sky view." +
+						"Ensure the antenna is outdoors if signal is weak or unavailable indoors.")
+				}
+			}
+		}
+	})
 }
 
 // ParseAndUpdate passes the provided message into the inner NmeaParser object, which parses the
@@ -205,4 +261,11 @@ func (g *CachedData) calculateCompassDegreeError(p1, p2 *geo.Point) float64 {
 	thetaRadians := math.Atan2(radius, adjacent)
 	thetaDegrees := utils.RadToDeg(thetaRadians)
 	return thetaDegrees
+}
+
+// Close shuts down the DataReader feeding this struct.
+func (g *CachedData) Close(ctx context.Context) error {
+	g.cancelFunc()
+	g.activeBackgroundWorkers.Wait()
+	return g.dev.Close()
 }
