+++ components/movementsensor/gpsutils/cachedData_test.go
@@ -1,4 +1,4 @@
-package rtkutils
+package gpsutils
 
 import (
 	"context"
@@ -6,6 +6,8 @@ import (
 
 	geo "github.com/kellydunn/golang-geo"
 	"go.viam.com/test"
+
+	"go.viam.com/rdk/logging"
 )
 
 var loc = geo.NewPoint(90, 1)
@@ -20,9 +22,20 @@ const (
 	fix        = 1
 )
 
+type mockDataReader struct{}
+
+func (d *mockDataReader) Messages() chan string {
+	return nil
+}
+
+func (d *mockDataReader) Close() error {
+	return nil
+}
+
 func TestReadingsSerial(t *testing.T) {
 	ctx := context.Background()
-	g := NewCachedData()
+	logger := logging.NewTestLogger(t)
+	g := NewCachedData(&mockDataReader{}, logger)
 	g.nmeaData = NmeaParser{
 		Location:   loc,
 		Alt:        alt,
